fix regression - loading config
