+++ react/public/index.html
@@ -11,6 +11,6 @@
       You need to enable JavaScript to run this app.
     </noscript>
     <div id="root" class="wrap container"></div>
-    <script type="text/javascript" src="/static/build/index.js"></script>
+    <script type="text/javascript" src="/static/build/index-bundle.dev.js"></script>
   </body>
 </html>
